+++ walking-beans-backend/src/main/java/walking_beans/walking_beans_backend/mapper/PaymentMapper.java
@@ -7,4 +7,7 @@ import walking_beans.walking_beans_backend.model.dto.Payments;
 public interface PaymentMapper {
     // 결제 정보 저장하기
     void insertPayments(Payments payments);
+
+    // 결제 정보 조회
+    Payments getPaymentByOrderId(long orderId);
 }
