+++ walking-beans-backend/src/main/java/walking_beans/walking_beans_backend/model/dto/Payments.java
@@ -15,5 +15,4 @@ public class Payments {
     private long orderId;
     private String paymentMethod;
     private String paymentStatus;
-
 }
