+++ walking-beans-frontend/src/pages/user/UserSuccessPage.jsx
@@ -83,7 +83,11 @@ const UserSuccessPage = () => {
                     storeId,
                     addressId,
                     orderRequests,
-                    cartList: cartItems
+                    cartList: cartItems,
+                    payments: {
+                        paymentMethod: "tossPay",
+                        paymentStatus: "완료"
+                    }
                 });
                 console.log("결제 승인 성공:", response.data);
                 alert("결제가 성공적으로 완료되었습니다!");
